backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0

pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=42.0.0
email-validator>=2.1.1

celery==5.3.6
redis==5.0.1
apscheduler==3.10.4

httpx==0.26.0
requests==2.31.0
aiofiles==23.2.1
websockets==12.0
tenacity==8.2.3
structlog==24.1.0
prometheus-client==0.19.0

pillow>=10.4.0
opencv-python-headless>=4.10.0

numpy>=2.1.0
pandas>=2.2.0
scikit-learn>=1.5.0

ultralytics>=8.3.0
torch>=2.3.0
torchvision>=0.18.0

shapely>=2.0.4
geopy==2.4.1

pymongo==4.6.1
motor==3.3.2

pytest==7.4.4
pytest-asyncio==0.23.3